requirement.txt
psycopg
python-dotenv>=1.0.0
openai>=1.3.0
google-generativeai>=0.3.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.65.0
requests>=2.31.0
urllib3>=2.0.0
datasets>=2.14.0
psycopg2
openreview-py
transformers
torch
accelerate
openpyxl
deepseek
matplotlib
seaborn
tiktoken
python-liquid